Refactor processing functions to improve data handling and add debug prints

+++ src/reco_plugin/_writer.py
@@ -17,5 +17,5 @@ def write_tiff(path: str, data: Any, meta: dict) -> list[str]:
         name = os.path.splitext(path)[0]
         path = f"{name}.tif"
 
-    tifffile.imwrite(path, data.astype(np.float32))
+    tifffile.imwrite(path, data)
     return [path]
